feat: 减少打开窗口，进一步提高效率
